+++ test/TestApp.AspNetCore/ActivityMiddleware.cs
@@ -28,17 +28,4 @@ public class ActivityMiddleware
             this.core.PostProcess(context);
         }
     }
-
-    public class ActivityMiddlewareCore
-    {
-        public virtual void PreProcess(HttpContext context)
-        {
-            // Do nothing
-        }
-
-        public virtual void PostProcess(HttpContext context)
-        {
-            // Do nothing
-        }
-    }
 }
